% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APD120A2.R
\name{APD120A2}
\alias{APD120A2}
\title{Power from Thorlabs APD120A2 given signal voltage}
\usage{
APD120A2(signal_mV, wavelength_nm = 550)
}
\arguments{
\item{signal_mV}{signal measured on lock-in amplifier in mV}

\item{wavelength_nm}{incident wavelength in nanometers}
}
\value{
Returns the incident power (in nW) on the photodetector
}
\description{
Power from Thorlabs Avalanche PhotoDetector Number APD120A2 for a given signal voltage
}
\examples{
APD120A2(10)

}
