% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_class.R
\docType{class}
\name{thorlabs_filter-class}
\alias{thorlabs_filter-class}
\alias{thorlabs_filter}
\title{Title}
\arguments{
\item{name}{The ThorLabs designation of the filter (for example FEL0550)}
}
\description{
Title
}

\examples{

}
