% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{thorlabs_filter.data}
\alias{thorlabs_filter.data}
\title{Get the raw transmission spectra data of a filter provided by Thorlabs}
\usage{
thorlabs_filter.data(name)
}
\arguments{
\item{name}{The ThorLabs designation of the filter (for example FEL0550)}
}
\value{
returns the provided transmission spectra from thorlabs as a dataframe
}
\description{
Get the raw transmission spectra data of a filter provided by Thorlabs
}
\examples{
thorlabs_filter.data("FEL0700")
}
