% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{thorlabs_filter.exists}
\alias{thorlabs_filter.exists}
\title{Check filter exists}
\usage{
thorlabs_filter.exists(name)
}
\arguments{
\item{name}{The ThorLabs designation of the filter (for example FEL0550)}
}
\description{
Check filter exists
}
\note{
This is an internal function for thor2, and not intended for use outside of it.
}
