% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{thorlabs_filter.import}
\alias{thorlabs_filter.import}
\title{Export filter to global enivroment}
\usage{
thorlabs_filter.import(name)
}
\arguments{
\item{name}{The ThorLabs designation of the filter (for example FEL0550)}
}
\value{
A spline function named after the filter.
Note if the thorlabs filter name contains a dash (for example, to designate bandwidth), this will be replaced in the function name as an underscore. i.e. FB340-10 becomes FB340_10
}
\description{
Exports the given filter as a spline function to the global enviroment, under it's own named function.
}
\examples{
thorlabs_filter.import("FES0650")
FES0650(500)

}
