% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{thorlabs_filter.plot}
\alias{thorlabs_filter.plot}
\title{Plot the Transmission Spectra of Thorlabs Filter}
\usage{
thorlabs_filter.plot(name, ...)
}
\arguments{
\item{name}{The ThorLabs designation of the filter (for example FEL0550)}

\item{...}{graphical arguments}
}
\value{
Plots to current device the provided transmission spectra (points) and the spline fit (line).
}
\description{
Plot the Transmission Spectra of Thorlabs Filter
}
\examples{
thorlabs_filter.plot("FEL0700")
}
