% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{thorlabs_filter}
\alias{thorlabs_filter}
\title{Transmission of ThorLabs Dielectric Filters}
\usage{
thorlabs_filter(name, wavelength_nm)
}
\arguments{
\item{name}{The ThorLabs designation of the filter (for example FEL0550)}

\item{wavelength_nm}{The wavelength(s) incident on the filter}
}
\value{
The fractional transmission through the filter
}
\description{
Transmission of ThorLabs Dielectric Filters
}
\examples{
thorlabs_filter("FEL0450", 200:700)

plot(200:700, thorlabs_filter("FES0550", 200:700), xlab = "wavelength (nm)", ylab = "T")

}
