% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{thorlabs_filter.spline}
\alias{thorlabs_filter.spline}
\title{Get spline function from Transmission data}
\usage{
thorlabs_filter.spline(name)
}
\arguments{
\item{name}{The ThorLabs designation of the filter (for example FEL0550)}
}
\value{
A spline function accepting wavelength_nm argument
}
\description{
Get spline function from Transmission data
}
\note{
This is an internal function for thor2, and not intended for use outside of it
}
\examples{
FEL0450 <- thorlabs_filter.spline("FEL0450")

}
