% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_class.R
\name{thorlabs_filter_import}
\alias{thorlabs_filter_import}
\title{Title}
\usage{
thorlabs_filter_import(name)
}
\arguments{
\item{name}{The ThorLabs designation of the filter (for example FEL0550)}
}
\description{
Title
}
\examples{
thorlabs_filter_import("FEL0650")
FEL0650$Transmission(wavelength_nm = 500)
}
