% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{thorlabs_filter.import}
\alias{thorlabs_filter.import}
\title{Export filter to global enivroment}
\usage{
thorlabs_filter.import(name)
}
\arguments{
\item{name}{The ThorLabs designation of the filter (for example FEL0550)}
}
\value{
a spline function named after the filter
}
\description{
Exports the given filter as a spline function to the global enviroment, under it's own named function.
}
\examples{
thorlabs_filter.import("FES0650")
FES0650(500)

}
